Consolidate duplicated WrappedStream test helpers to Common sources
