+++ src/libraries/Common/tests/System/IO/WrappedStream.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.IO;
-
-namespace System.Formats.Tar
+namespace System.IO
 {
     public class WrappedStream : Stream
     {
